massive reorg to compiler, coverage, build
